+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 
 - **Breaking Change**: clear exit status when enter catch block
 - cancel runtime glob expansion by SIGINT
+- cancel runtime code completion by SIGINT
 - show sub-shell level when handle uncaught exception
 
 #### Builtin